feat: add navbar
